move buildParallelBlockMappings function to funnel utils
