+++ packages/engine/paima-funnel/src/utils.ts
@@ -95,3 +95,40 @@ export async function findBlockByTimestamp(
 
   return low;
 }
+
+// deterministically assigns to every block in blockData, a block in chainData.
+// this is the block that is used for the data merge.
+// additionally returns the inverse mapping.
+export function buildParallelBlockMappings(
+  applyDelay: (ts: number) => number,
+  chainData: ChainData[],
+  blockData: [number, { blockNumber: number }][]
+): {
+  parallelToMainchainBlockHeightMapping: { [blockNumber: number]: number };
+  mainchainToParallelBlockHeightMapping: { [blockNumber: number]: number };
+} {
+  const parallelToMainchainBlockHeightMapping: { [blockNumber: number]: number } = {};
+  const mainchainToParallelBlockHeightMapping: { [blockNumber: number]: number } = {};
+
+  let currIndex = 0;
+
+  for (const block of blockData) {
+    while (currIndex < chainData.length) {
+      if (applyDelay(chainData[currIndex].timestamp) >= block[0]) {
+        parallelToMainchainBlockHeightMapping[block[1].blockNumber] =
+          chainData[currIndex].blockNumber;
+
+        mainchainToParallelBlockHeightMapping[chainData[currIndex].blockNumber] =
+          block[1].blockNumber;
+        break;
+      } else {
+        currIndex++;
+      }
+    }
+  }
+
+  return {
+    parallelToMainchainBlockHeightMapping: parallelToMainchainBlockHeightMapping,
+    mainchainToParallelBlockHeightMapping: mainchainToParallelBlockHeightMapping,
+  };
+}
